+++ dipimage/resample.m
@@ -1,15 +1,11 @@
-%ROTATION   Rotate an image within an orthogonal plane
+%RESAMPLE   Shift and a scale an image using interpolation
 %
 % SYNOPSIS:
-%  image_out = rotation(image_in, angle, interpolation_method, boundary_condition) % (for 2D images only)
-%  image_out = rotation(image_in, angle, axis, interpolation_method, boundary_condition) % (for 3D images only)
-%  image_out = rotation(image_in, angle, dimension1, dimension2, interpolation_method, boundary_condition) % (the general case)
+%  image_out = resample(image_in, zoom, shift, interpolation_method, boundary_condition)
 %
 % PARAMETERS:
-%  angle: Angle (in radian) to rotate over
-%  axis: Axis to rotate around (for 2D or 3D images only, for a 2D image it has to be 3).
-%  dimension1: First dimension defining the orthogonal plane
-%  dimension2: Second dimension defining the orthogonal plane
+%  zoom: array containing a zoom for each dimension
+%  shift: array containing a shift for each dimension
 %  interpolation_method: One of the following strings:
 %                        - '3-cubic' or '': third order cubic spline interpolation
 %                        - '4-cubic': fourth order cubic spline interpolation
@@ -25,25 +21,24 @@
 %  boundary_condition: Defines how the boundary of the image is handled.
 %                      See HELP BOUNDARY_CONDITION
 %
+% MATHEMATICAL OPERATION:
+%  With ZOOM and SHIFT as vectors, and POS the position vector:
+%     image_out(pos) = image_in(pos/zoom-shift);
+%
 % DEFAULTS:
-%  dimension1 = 1
-%  dimension2 = dimension1 + 1  (that is: axis = 3)
-%  interpolation_method = '3-cubic'
-%  boundary_condition = 'add zeros'
+%  zoom = 2
+%  shift = 0
+%  interpolation_method = ''
+%  boundary_condition = ''
 %
-% The image is rotated around its center, using three skews.
-%
-% NOTES:
-%  Sign of the 2D rotation: implementation in the mathmetical sense, but
-%  note the y-axis is positive downwards! Thus: left turning has negative
-%  sign, and right positive.
-%
-%  For exact rotations of multiples of 90 degrees, use the rot90 method
-%  instead. Note that rot90 reverses the sign of the angle as compared to
-%  this function.
+% EXAMPLE:
+%  a = readim;
+%  b = resample(a,[.5,1],[-40,0])
+%  c = readim('chromo3d');
+%  d = resample(c,[1.2,1,2],[0,30,0])
 %
 % DIPlib:
-%  This function calls the DIPlib function dip::Rotation.
+%  This function calls the DIPlib function dip::Resampling.
 
 % (c)2017, Cris Luengo.
 % Based on original DIPlib code: (c)1995-2014, Delft University of Technology.